Added a simple Makefile for the sample code.

+++ customizable-printf/sample-print.c
@@ -1,4 +1,4 @@
-// Copyright 2020 Steven Buytaert.
+// Copyright 2020, 2023 Steven Buytaert.
 
 #include <stdio.h>
 #include <assert.h>
@@ -62,7 +62,7 @@ void myprintf(const char * fmt, ...) {                      // The only exported
   cux_vprintf(& Fmt, fmt, ap);
   va_end(ap);
 
-  ssize_t fu = write(1, buf, Fmt.cursor - buf);
+  ssize_t fu = write(1, buf, (size_t)(Fmt.cursor - buf));
   
   (void) fu;
